refactor: extract @covel/settings from the shared kitchen-sink package
Audit root-cause 5: @covel/shared bundled the SettingsStore + its platform backends
(browser localStorage, Electron-IPC json-file), forcing every pure-types consumer to pull in
browser/Electron code. Moved settings/store/types/backends into a new @covel/settings package
(git-tracked as renames); shared is now platform-code-free. Updated the 14 apps/web importers
+ added the missing zod dep. Dep flow: shared <- settings <- web. json-render-presets + env
groups stay in shared (no platform code). Doc-sync CLAUDE.md (15->16 packages).
Verified: pnpm install + build 3/3 + lint 18/18 + test 35/35.
