+++ apps/web/package.json
@@ -16,6 +16,7 @@
     "test:watch": "vitest"
   },
   "dependencies": {
+    "@covel/settings": "workspace:*",
     "@covel/shared": "workspace:*",
     "@covel/store": "workspace:*",
     "@json-render/core": "^0.18.0",
