+++ apps/web/src/theme-system/settings.ts
@@ -1,4 +1,4 @@
-import type { SettingsStoreApi } from "@covel/shared";
+import type { SettingsStoreApi } from "@covel/settings";
 import { primeThemeRegistry } from "./registry.js";
 
 export function registerThemeSettings(store: SettingsStoreApi): void {
