+++ packages/settings/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@covel/settings",
+  "version": "0.0.7",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "lint": "tsc --noEmit",
+    "clean": "node ../../scripts/remove-paths.mjs dist"
+  },
+  "dependencies": {
+    "@covel/shared": "workspace:*",
+    "zod": "^4.3.6"
+  },
+  "devDependencies": {
+    "typescript": "^6.0.3",
+    "vitest": "^4.1.4"
+  }
+}
